Hide sections with no options
Keep references to the headers of each section.
If the section does not has any visible options,
the section is also not visible.
During initialization, we do not create DOM elements
for sections with no data elements.
When we select all options of a section, this section
disappears.
When we deselect one option that belongs to a previously hidden
section, the section reappears.

Fixes: #120.